remove install via pip
