+++ src/main.rs
@@ -2,21 +2,18 @@ use camino::{Utf8Path, Utf8PathBuf};
 use eyre::{ensure, eyre, WrapErr};
 use fs_err::File;
 use itertools::Itertools;
-use python_requirements::Requirement;
 use python_wheel::{RecordEntry, WheelRecord};
 use rand::Rng;
 use sha2::{Digest, Sha256};
 use std::collections::HashSet;
 use std::convert::TryFrom;
 use std::convert::TryInto;
-use std::fmt::Write;
 use std::io::Seek;
 use std::{collections::HashMap, io::BufReader, path::Path as StdPath};
 use structopt::StructOpt;
 
 mod internal_store;
 mod python_detection;
-mod python_requirements;
 mod python_wheel;
 
 use fs_err::PathExt;
@@ -56,18 +53,6 @@ enum Command {
         python: String,
         #[structopt(long)]
         without_pip_shim: bool,
-    },
-    /// Add dependency to virtpy
-    Add {
-        requirements: PathBuf,
-        #[structopt(long)]
-        virtpy_path: Option<PathBuf>,
-    },
-    /// Remove dependency from virtpy
-    Remove {
-        distributions: Vec<String>,
-        #[structopt(long)]
-        virtpy_path: Option<PathBuf>,
     },
     /// Install executable package into an isolated virtpy
     Install {
@@ -123,7 +108,15 @@ enum InternalStoreCmd {
 
 #[derive(StructOpt)]
 enum InternalUseOnly {
-    AddFromFile { virtpy: PathBuf, file: PathBuf },
+    AddFromFile {
+        virtpy: PathBuf,
+        file: PathBuf,
+    },
+    /// Remove distribution from virtpy
+    Remove {
+        virtpy: PathBuf,
+        distributions: Vec<String>,
+    },
 }
 
 #[derive(StructOpt)]
@@ -511,42 +504,6 @@ impl std::ops::DerefMut for FileLockGuard {
     }
 }
 
-fn serialize_requirements_txt(reqs: &[Requirement]) -> String {
-    let mut output = String::new();
-    for req in reqs {
-        let _ = write!(&mut output, "{}=={}", req.name, req.version);
-        if let Some(marker) = req.marker.as_ref() {
-            let _ = write!(&mut output, "; {}", marker);
-        }
-        let _ = writeln!(&mut output, " \\");
-        let hashes = req
-            .available_hashes
-            .iter()
-            .map(|hash| format!("    --hash={}", hash.0.replace("=", ":")))
-            .collect::<Vec<_>>();
-        let _ = writeln!(&mut output, "{}", hashes.join(" \\\n"));
-    }
-    output
-}
-
-fn copy_directory(from: impl AsRef<StdPath>, to: impl AsRef<StdPath>, use_move: bool) {
-    _copy_directory(from.as_ref(), to.as_ref(), use_move)
-}
-
-fn _copy_directory(from: &StdPath, to: &StdPath, use_move: bool) {
-    for dir_entry in walkdir::WalkDir::new(from) {
-        let dir_entry = dir_entry.unwrap();
-        let path = dir_entry.path();
-        let subpath = path.strip_prefix(from).unwrap();
-        let target_path = to.join(&subpath);
-        if dir_entry.file_type().is_dir() {
-            fs_err::create_dir(target_path).unwrap();
-        } else {
-            move_file(path, &target_path, use_move).unwrap();
-        }
-    }
-}
-
 fn is_path_of_executable(path: &Utf8Path) -> bool {
     path.starts_with("bin") || path.starts_with("Scripts")
 }
@@ -711,81 +668,6 @@ fn _move_file(src: &StdPath, dst: &StdPath, use_move: bool) -> std::io::Result<(
     }
 }
 
-fn register_distribution_files(
-    proj_dirs: &ProjectDirs,
-    install_folder: &Path,
-    distribution: &Distribution,
-    stored_distributions: &mut HashMap<DistributionHash, StoredDistribution>,
-    options: crate::Options,
-) -> EResult<()> {
-    let dist_info_foldername = distribution.dist_info_name();
-    let src_dist_info = install_folder.join(&dist_info_foldername);
-
-    let stored_distrib = StoredDistribution {
-        distribution: distribution.clone(),
-        installed_via: StoredDistributionType::FromPip,
-    };
-    let dst_dist_info = proj_dirs
-        .dist_infos()
-        .join(&stored_distrib.distribution.as_csv());
-
-    // let use_move = can_move_files(&proj_dirs.package_files(), install_folder).unwrap_or(false);
-    let use_move = true;
-
-    if dst_dist_info.exists() {
-        // add it here, because it may have been installed by a different
-        // python version. In that case, the current python version's list
-        // may be missing this distribution.
-        stored_distributions.insert(distribution.sha.clone(), stored_distrib);
-        return Ok(());
-    }
-    if options.verbose >= 1 {
-        println!(
-            "Adding {} {} to central store.",
-            distribution.name, distribution.version
-        );
-    }
-
-    for file in records(&src_dist_info.join("RECORD"))
-        .wrap_err("couldn't find dist-info/RECORD")?
-        .map(Result::unwrap)
-    {
-        // Sanity check. We're not caching compiled code so pip is told not to compile python code.
-        // If this folder exists, something went wrong.
-        debug_assert!(file.path.iter().all(|part| part != "__pycache__"));
-
-        let path = remove_leading_parent_dirs(&file.path).unwrap_or_else(std::convert::identity);
-        if is_path_of_executable(path) {
-            // executables generated by pip depend either on the global python
-            // or the venv they were generated for. They can't be symlinked into a virtpy
-            // but have to be generated on demand.
-            continue;
-        }
-
-        debug_assert_ne!(file.hash, FileHash("".to_owned()));
-
-        let src = install_folder.join(path);
-        let dest = proj_dirs.package_files().join(file.hash.0);
-        if options.verbose >= 2 {
-            println!("    copying {} to {}", src, dest);
-        }
-
-        let res = move_file(&src, &dest, use_move);
-        match &res {
-            Err(err) if is_not_found(err) => {
-                print_error_missing_file_in_record(&distribution, file.path.as_ref())
-            }
-            _ => {
-                res.unwrap();
-            }
-        };
-    }
-
-    copy_directory(&src_dist_info, &dst_dist_info, use_move);
-    stored_distributions.insert(distribution.sha.clone(), stored_distrib);
-    Ok(())
-}
-
 fn register_distribution_files_of_wheel(
     proj_dirs: &ProjectDirs,
     install_folder: &Path,
@@ -895,7 +777,7 @@ fn records(record: &Path) -> csv::Result<impl Iterator<Item = csv::Result<Record
 fn install_and_register_distribution_from_file(
     proj_dirs: &ProjectDirs,
     distrib_path: &Path,
-    requirement: Requirement,
+    distrib: Distribution,
     python_version: PythonVersion,
     options: Options,
 ) -> EResult<()> {
@@ -913,12 +795,6 @@ fn install_and_register_distribution_from_file(
         &distrib_path,
         tmp_dir.path().try_into().expect(INVALID_UTF8_PATH),
     )?;
-
-    let distrib = Distribution {
-        name: requirement.name,
-        version: requirement.version,
-        sha: requirement.available_hashes.into_iter().next().unwrap(),
-    };
 
     register_new_distribution(options, distrib, proj_dirs, python_version, tmp_dir)?;
 
@@ -975,109 +851,6 @@ fn _convert_to_wheel(
     }
 }
 
-fn install_and_register_distributions(
-    python_path: &Path,
-    proj_dirs: &ProjectDirs,
-    distribs: &[Requirement],
-    python_version: PythonVersion,
-    options: Options,
-) -> EResult<()> {
-    if options.verbose >= 1 {
-        println!("Adding {} new distributions", distribs.len());
-    }
-    if distribs.is_empty() {
-        return Ok(());
-    }
-
-    let tmp_dir = tempdir::TempDir::new_in(proj_dirs.tmp(), "virtpy")?;
-    let tmp_requirements = tmp_dir.as_ref().join("__tmp_requirements.txt");
-    let reqs = serialize_requirements_txt(distribs);
-    fs_err::write(&tmp_requirements, reqs)?;
-    let output = std::process::Command::new(python_path)
-        .args(&["-m", "pip", "install", "--no-deps", "--no-compile", "-r"])
-        .arg(&tmp_requirements)
-        .arg("-t")
-        .arg(tmp_dir.as_ref())
-        .arg("-v")
-        .output()?;
-    if !output.status.success() {
-        panic!(
-            "pip error:\n{}",
-            std::str::from_utf8(&output.stderr).unwrap()
-        );
-    }
-
-    let pip_log = String::from_utf8(output.stdout)?;
-
-    let new_distribs = newly_installed_distributions(&pip_log);
-    register_new_distributions(
-        options,
-        new_distribs,
-        distribs.len(),
-        proj_dirs,
-        pip_log,
-        python_version,
-        tmp_dir,
-    )?;
-
-    Ok(())
-}
-
-fn register_new_distributions(
-    options: Options,
-    new_distribs: Vec<Distribution>,
-    n_distribs_requested: usize,
-    proj_dirs: &ProjectDirs,
-    pip_log: String,
-    python_version: PythonVersion,
-    tmp_dir: tempdir::TempDir,
-) -> EResult<()> {
-    if options.verbose >= 1 {
-        if new_distribs.len() != n_distribs_requested {
-            // either an error or a sign that the filters in new_dependencies()
-            // need to be improved
-            println!(
-                "Only found {} of {} distributions",
-                new_distribs.len(),
-                n_distribs_requested
-            );
-
-            let _ = fs_err::write(proj_dirs.data().join("pip.log"), pip_log);
-        }
-    }
-    if options.verbose >= 2 {
-        for distrib in new_distribs.iter() {
-            println!(
-                "    New distribution: {}=={}, {}",
-                distrib.name, distrib.version, distrib.sha.0
-            );
-        }
-    }
-    let mut all_stored_distributions = StoredDistributions::load(proj_dirs)?;
-    let stored_distributions = all_stored_distributions
-        .0
-        .entry(python_version.as_string_without_patch())
-        .or_default();
-    for distrib in new_distribs {
-        register_distribution_files(
-            proj_dirs,
-            tmp_dir.path().try_into().expect(INVALID_UTF8_PATH),
-            &distrib,
-            stored_distributions,
-            options,
-        )
-        .wrap_err_with(|| {
-            eyre!(
-                "failed to add distribution files for {} {}",
-                distrib.name,
-                distrib.version
-            )
-        })?;
-    }
-    all_stored_distributions.save()?;
-    Ok(())
-}
-
 // Usable only for our own installation from wheel files
 fn register_new_distribution(
     options: Options,
@@ -1113,35 +886,6 @@ fn register_new_distribution(
     })?;
     all_stored_distributions.save()?;
     Ok(())
-}
-
-fn new_dependencies(
-    requirements: &[Requirement],
-    proj_dirs: &ProjectDirs,
-    python_version: PythonVersion,
-) -> EResult<Vec<Requirement>> {
-    let stored_distributions = StoredDistributions::load(proj_dirs)?;
-    let existing_deps = match stored_distributions
-        .0
-        .get(&python_version.as_string_without_patch())
-    {
-        Some(deps) => deps,
-        None => return Ok(requirements.to_owned()),
-    };
-
-    Ok(requirements
-        .iter()
-        .filter(|req| {
-            (req.marker
-                .as_ref()
-                .map_or(true, |cond| cond.matches_system()))
-                && !req
-                    .available_hashes
-                    .iter()
-                    .any(|hash| existing_deps.contains_key(hash))
-        })
-        .cloned()
-        .collect::<Vec<_>>())
 }
 
 fn wheel_is_already_registered(
@@ -1326,34 +1070,6 @@ fn main() -> EResult<()> {
     proj_dirs.create_dirs()?;
 
     match opt.cmd {
-        Command::Add {
-            requirements,
-            virtpy_path,
-        } => {
-            fn add_requirements(
-                proj_dirs: &ProjectDirs,
-                virtpy_path: Option<PathBuf>,
-                options: Options,
-                requirements: PathBuf,
-            ) -> EResult<()> {
-                let virtpy = CheckedVirtpy::new(path_to_virtpy(&virtpy_path))?;
-                let requirements = fs_err::read_to_string(requirements)?;
-                let requirements = python_requirements::read_requirements_txt(&requirements);
-
-                virtpy_add_dependencies(&proj_dirs, &virtpy, requirements, options)?;
-                Ok(())
-            }
-
-            add_requirements(&proj_dirs, virtpy_path, options, requirements)
-                .wrap_err("failed to add requirements")?;
-        }
-        Command::Remove {
-            distributions,
-            virtpy_path,
-        } => {
-            let virtpy = CheckedVirtpy::new(path_to_virtpy(&virtpy_path))?;
-            virtpy_remove_dependencies(&virtpy, distributions.into_iter().collect())?;
-        }
         Command::New {
             path,
             python,
@@ -1420,6 +1136,13 @@ fn main() -> EResult<()> {
         Command::InternalUseOnly(InternalUseOnly::AddFromFile { virtpy, file }) => {
             let virtpy = CheckedVirtpy::new(&virtpy)?;
             virtpy_add_dependency_from_file(&proj_dirs, &virtpy, &file, options)?;
+        }
+        Command::InternalUseOnly(InternalUseOnly::Remove {
+            virtpy,
+            distributions,
+        }) => {
+            let virtpy = CheckedVirtpy::new(path_to_virtpy(&Some(virtpy)))?;
+            virtpy_remove_dependencies(&virtpy, distributions.into_iter().collect())?;
         }
     }
 
@@ -1673,30 +1396,6 @@ fn _hash_of_reader_sha256(mut reader: impl std::io::Read) -> impl AsRef<[u8]> {
     hasher.finalize()
 }
 
-fn virtpy_add_dependencies(
-    proj_dirs: &ProjectDirs,
-    virtpy: &CheckedVirtpy,
-    requirements: Vec<Requirement>,
-    //python_version: PythonVersion,
-    options: Options,
-) -> EResult<()> {
-    let new_deps = new_dependencies(&requirements, proj_dirs, virtpy.python_version)?;
-
-    // The virtpy doesn't contain pip so get the appropriate global python
-    let python_path = virtpy.global_python()?;
-
-    install_and_register_distributions(
-        &python_path,
-        proj_dirs,
-        &new_deps,
-        virtpy.python_version,
-        options,
-    )?;
-
-    link_requirements_into_virtpy(proj_dirs, virtpy, requirements, options)
-        .wrap_err("failed to add packages to virtpy")
-}
-
 fn virtpy_add_dependency_from_file(
     proj_dirs: &ProjectDirs,
     virtpy: &CheckedVirtpy,
@@ -1704,20 +1403,20 @@ fn virtpy_add_dependency_from_file(
     options: Options,
 ) -> EResult<()> {
     let file_hash = DistributionHash::from_file(file);
-    let requirement =
-        Requirement::from_filename(file.file_name().unwrap(), file_hash.clone()).unwrap();
+    let distribution =
+        Distribution::from_filename(file.file_name().unwrap(), file_hash.clone()).unwrap();
 
     if !wheel_is_already_registered(file_hash.clone(), proj_dirs, virtpy.python_version)? {
         install_and_register_distribution_from_file(
             proj_dirs,
             file,
-            requirement.clone(),
+            distribution.clone(),
             virtpy.python_version,
             options,
         )?;
     }
 
-    link_requirements_into_virtpy(proj_dirs, virtpy, vec![requirement], options)
+    link_requirements_into_virtpy(proj_dirs, virtpy, vec![distribution], options)
         .wrap_err("failed to add packages to virtpy")
 }
 
@@ -2151,6 +1850,7 @@ impl CheckedVirtpy {
 
     // Returns the path of the python installation on which this
     // this virtpy builds
+    #[allow(unused)]
     fn global_python(&self) -> EResult<PathBuf> {
         #[cfg(unix)]
         {
@@ -2209,7 +1909,7 @@ impl CheckedVirtpy {
 fn link_requirements_into_virtpy(
     proj_dirs: &ProjectDirs,
     virtpy: &CheckedVirtpy,
-    mut requirements: Vec<Requirement>,
+    distributions: Vec<Distribution>,
     options: Options,
 ) -> EResult<()> {
     // Link files into the backing virtpy so that when new top-level directories are
@@ -2217,26 +1917,15 @@ fn link_requirements_into_virtpy(
     // Symlinks for the new dirs are generated after all the files have been liked in.
     let site_packages = virtpy.virtpy_backing().site_packages();
 
-    requirements.retain(|req| {
-        req.marker
-            .as_ref()
-            .map_or(true, |cond| cond.matches_system())
-    });
-    let requirements = requirements;
-
     let stored_distributions = StoredDistributions::load(proj_dirs)?;
     let existing_deps = stored_distributions
         .0
         .get(&virtpy.python_version.as_string_without_patch())
         .cloned()
         .unwrap_or_default();
-    for distribution in requirements {
+    for distribution in distributions {
         // find compatible hash
-        let stored_distrib = match distribution
-            .available_hashes
-            .iter()
-            .find_map(|hash| existing_deps.get(&hash))
-        {
+        let stored_distrib = match existing_deps.get(&distribution.sha) {
             Some(stored_distrib) => stored_distrib,
             None => {
                 // return Err(format!(
@@ -2245,16 +1934,9 @@ fn link_requirements_into_virtpy(
                 // )
                 // .into());
                 println!(
-                    "failed to find dist_info for distribution: {} {} {}",
-                    distribution.name,
-                    distribution.version,
-                    distribution
-                        .marker
-                        .map_or(String::new(), |m| format!(", {}", m))
+                    "failed to find dist_info for distribution: {} {}",
+                    distribution.name, distribution.version,
                 );
-                if options.verbose >= 2 {
-                    println!("available_hashes: {:#?}", distribution.available_hashes);
-                }
                 continue;
             }
         };
@@ -2585,6 +2267,20 @@ struct Distribution {
 }
 
 impl Distribution {
+    pub fn from_filename(filename: &str, hash: DistributionHash) -> EResult<Self> {
+        // TODO: use a better parser
+        let (_, name, version, _, _) =
+            lazy_regex::regex_captures!(r"^([\w\d]+)-(\d+(\.\d+)*)(\.tar\.gz|.*\.whl)", filename)
+                .unwrap();
+
+        Ok(Self {
+            name: name.to_owned(),
+            version: version.to_owned(),
+            sha: hash,
+        })
+        //RequirementsTxtParser::parse(Rule::wheel_name, wheel_name).unwrap();
+    }
+
     fn from_store_name(store_name: &str) -> Self {
         let (_, name, version, hash) =
             lazy_regex::regex_captures!(r"([^,]+),([^,]+),([^,]+)", store_name).unwrap();
@@ -2614,35 +2310,6 @@ impl Distribution {
         format!("{}-{}.data", self.name, self.version)
     }
 }
-
-fn newly_installed_distributions(pip_log: &str) -> Vec<Distribution> {
-    let mut installed_distribs = Vec::new();
-
-    let install_url_pattern = lazy_regex::regex!(
-        r"Added ([\w_-]+)==(.*) from (https://[^\s]+)/([\w_]+)-[\w_\-\.]+#(sha256=[0-9a-fA-F]{64})"
-    );
-
-    for line in pip_log.lines() {
-        if let Some(install_captures) = install_url_pattern.captures(line) {
-            let get = |idx| install_captures.get(idx).unwrap().as_str().to_owned();
-            // false name, may not have right case
-            //let name = get(1);
-            let version = get(2);
-            //let url = get(3);
-            let name = get(4);
-            let sha = DistributionHash(get(5));
-
-            //installed_distribs.push((url, distribution, version));
-            installed_distribs.push(Distribution { version, sha, name })
-        } else if line.contains("Added ") {
-            // The regex should have matched.
-            panic!("2: {}", line);
-        }
-    }
-
-    installed_distribs
-}
-
 #[cfg(test)]
 mod test {
     use crate::python_detection::detect;
@@ -2693,81 +2360,6 @@ mod test {
             assert_eq!(proj_dirs.executables().read_dir().unwrap().count(), 0);
         }
         Ok(())
-    }
-
-    #[test]
-    fn test_pip_log_parsing() {
-        let text = include_str!("../test_files/pip.log");
-        let distribs = newly_installed_distributions(text);
-        assert_eq!(
-            distribs,
-            &[
-                Distribution {
-                    name: "astroid".into(),
-                    version: "2.4.2".into(),
-                    sha: DistributionHash(
-                        "sha256=bc58d83eb610252fd8de6363e39d4f1d0619c894b0ed24603b881c02e64c7386"
-                            .into()
-                    )
-                },
-                Distribution {
-                    name: "isort".into(),
-                    version: "4.3.21".into(),
-                    sha: DistributionHash(
-                        "sha256=6e811fcb295968434526407adb8796944f1988c5b65e8139058f2014cbe100fd"
-                            .into()
-                    )
-                },
-                Distribution {
-                    name: "lazy_object_proxy".into(),
-                    version: "1.4.3".into(),
-                    sha: DistributionHash(
-                        "sha256=a6ae12d08c0bf9909ce12385803a543bfe99b95fe01e752536a60af2b7797c62"
-                            .into()
-                    )
-                },
-                Distribution {
-                    name: "mccabe".into(),
-                    version: "0.6.1".into(),
-                    sha: DistributionHash(
-                        "sha256=ab8a6258860da4b6677da4bd2fe5dc2c659cff31b3ee4f7f5d64e79735b80d42"
-                            .into()
-                    )
-                },
-                Distribution {
-                    name: "pylint".into(),
-                    version: "2.5.3".into(),
-                    sha: DistributionHash(
-                        "sha256=d0ece7d223fe422088b0e8f13fa0a1e8eb745ebffcb8ed53d3e95394b6101a1c"
-                            .into()
-                    )
-                },
-                Distribution {
-                    name: "six".into(),
-                    version: "1.15.0".into(),
-                    sha: DistributionHash(
-                        "sha256=8b74bedcbbbaca38ff6d7491d76f2b06b3592611af620f8426e82dddb04a5ced"
-                            .into()
-                    )
-                },
-                Distribution {
-                    name: "toml".into(),
-                    version: "0.10.1".into(),
-                    sha: DistributionHash(
-                        "sha256=bda89d5935c2eac546d648028b9901107a595863cb36bae0c73ac804a9b4ce88"
-                            .into()
-                    )
-                },
-                Distribution {
-                    name: "wrapt".into(),
-                    version: "1.12.1".into(),
-                    sha: DistributionHash(
-                        "sha256=b62ffa81fb85f4332a4f609cab4ac40709470da05643a082ec1eb88e6d9b97d7"
-                            .into()
-                    )
-                }
-            ]
-        );
     }
 
     #[test]
@@ -2881,5 +2473,47 @@ mod test {
         case("/c0/c1/a0/a1/", "/c0/c1/b0/b1/b2", "../../b0/b1/b2");
         case("/c0/c1/a0/a1/a2", "/c0/c1/b0/b1/b2", "../../../b0/b1/b2");
         case("/c0/c1", "/c0/c1", "");
+    }
+
+    #[test]
+    fn test_parse_wheel_name_into_distribution() {
+        #[rustfmt::skip]
+        let filenames_and_output = [
+            ("astroid-2.4.2-py3-none-any.whl", ("astroid", "2.4.2")),
+            ("async_generator-1.10-py3-none-any.whl", ("async_generator", "1.10")),
+            ("attrs-19.3.0-py2.py3-none-any.whl", ("attrs", "19.3.0")),
+            ("click-7.1.2-py2.py3-none-any.whl", ("click", "7.1.2")),
+            ("idna-3.1-py3-none-any.whl", ("idna", "3.1")),
+            ("isort-4.3.21-py2.py3-none-any.whl", ("isort", "4.3.21")),
+            ("lazy_object_proxy-1.4.3-cp38-cp38-manylinux1_x86_64.whl", ("lazy_object_proxy", "1.4.3")),
+            ("mccabe-0.6.1-py2.py3-none-any.whl", ("mccabe", "0.6.1")),
+            ("more_itertools-8.4.0-py3-none-any.whl", ("more_itertools", "8.4.0")),
+            ("mypy-0.782-cp38-cp38-manylinux1_x86_64.whl", ("mypy", "0.782")),
+            ("mypy_extensions-0.4.3-py2.py3-none-any.whl", ("mypy_extensions", "0.4.3")),
+            ("outcome-1.1.0-py2.py3-none-any.whl", ("outcome", "1.1.0")),
+            ("packaging-20.4-py2.py3-none-any.whl", ("packaging", "20.4")),
+            ("pluggy-0.13.1-py2.py3-none-any.whl", ("pluggy", "0.13.1")),
+            ("py-1.8.2-py2.py3-none-any.whl", ("py", "1.8.2")),
+            ("pylint-2.5.3-py3-none-any.whl", ("pylint", "2.5.3")),
+            ("pyparsing-2.4.7-py2.py3-none-any.whl", ("pyparsing", "2.4.7")),
+            ("pytest-5.4.3-py3-none-any.whl", ("pytest", "5.4.3")),
+            ("six-1.15.0-py2.py3-none-any.whl", ("six", "1.15.0")),
+            ("sniffio-1.2.0-py3-none-any.whl", ("sniffio", "1.2.0")),
+            ("sortedcontainers-2.4.0-py2.py3-none-any.whl", ("sortedcontainers", "2.4.0")),
+            ("toml-0.10.1-py2.py3-none-any.whl", ("toml", "0.10.1")),
+            ("trio-0.18.0-py3-none-any.whl", ("trio", "0.18.0")),
+            ("typed_ast-1.4.1-cp38-cp38-manylinux1_x86_64.whl", ("typed_ast", "1.4.1")),
+            ("typing_extensions-3.7.4.2-py3-none-any.whl", ("typing_extensions", "3.7.4.2")),
+            ("wcwidth-0.2.5-py2.py3-none-any.whl", ("wcwidth", "0.2.5")),
+            ("wrapt-1.12.1.tar.gz", ("wrapt", "1.12.1")),
+        ];
+
+        for &(filename, (distrib_name, version)) in filenames_and_output.iter() {
+            let req =
+                crate::Distribution::from_filename(filename, crate::DistributionHash("".into()))
+                    .unwrap();
+            assert_eq!(req.name, distrib_name);
+            assert_eq!(req.version, version);
+        }
     }
 }
